refactor(tools): extract research domain into src/tools/research.py
